+++ lib/whois/record/parser/whois.nic.uk.rb
@@ -43,6 +43,8 @@ module Whois
                 :registered
               when "renewal request being processed."
                 :registered
+              when "no registration status listed."
+                :reserved
               # NEWSTATUS (redemption?)
               when "renewal required."
                 :registered
@@ -66,24 +68,25 @@ module Whois
 
 
         property_supported :created_on do
-          if content_for_scanner =~ /\s+Registered on:\s+(.*)\n/
+          if content_for_scanner =~ /\s+Registered on:\s+(.+)\n/
             Time.parse($1)
           end
         end
 
         property_supported :updated_on do
-          if content_for_scanner =~ /\s+Last updated:\s+(.*)\n/
+          if content_for_scanner =~ /\s+Last updated:\s+(.+)\n/
             Time.parse($1)
           end
         end
 
         property_supported :expires_on do
-          if content_for_scanner =~ /\s+Renewal date:\s+(.*)\n/
+          if content_for_scanner =~ /\s+Renewal date:\s+(.+)\n/
             Time.parse($1)
           end
         end
 
 
+        # @see http://www.nic.uk/other/whois/instruct/
         property_supported :registrar do
           if content_for_scanner =~ /Registrar:\n(.+) \[Tag = (.+)\]\n\s*URL: (.+)\n/
             name, id, url = $1.strip, $2.strip, $3.strip
@@ -91,21 +94,28 @@ module Whois
 
             Record::Registrar.new(
               :id           => id,
-              :url          => url,
               :name         => (name || org),
-              :organization => org
+              :organization => org,
+              :url          => url
+            )
+          elsif content_for_scanner =~ /This domain is registered directly with Nominet/
+            Record::Registrar.new(
+              :id           => nil,
+              :name         => "Nominet",
+              :organization => "Nominet UK",
+              :url          => "http://www.nic.uk/",
             )
           end
         end
 
 
         property_supported :registrant_contacts do
-          if address_info = content_for_scanner[/Registrant's address:\n(.+)\n\n/m, 1]
-            address, city, state, zip, country = address_info.split("\n").map(&:strip)
+          if content_for_scanner =~ /Registrant's address:\n((.+\n)+)\n/
+            *address, city, state, zip, country = *$1.split("\n").map(&:strip)
             Record::Contact.new(
               :type => Record::Contact::TYPE_REGISTRANT,
               :name => content_for_scanner[/Registrant:\n\s*(.+)\n/, 1],
-              :address => address,
+              :address => address.join("\n"),
               :city => city,
               :state => state,
               :zip => zip,